Add qlog as agenda item
I will post an update on this on the mailing list in the coming days as well. 

With draft-02, the qlog project is currently in a state that it is difficult to have meaningful progression without added community involvement. This could take many shapes (adoption by the wg, design team, separate IETF BOF/WG, an (informal) interim, simply more people interacting with the personal drafts, ...) and I think it's useful to get an idea of what the broader QUIC community thinks about this at this point.